Remove Frontend Testing Job from CI/CD Pipeline
This commit removes the frontend testing job from the GitHub Actions workflow in the Foodgram project. Changes include:

- The entire `frontend_tests` job, which was responsible for setting up Node.js, installing dependencies, and running tests for the frontend, has been removed from the workflow.
- The dependency on `frontend_tests` in the `build_frontend_and_push_to_docker_hub` job has also been eliminated.

This alteration simplifies the CI/CD pipeline by bypassing frontend tests. This decision may be based on a strategic change in testing methodology or a temporary measure to expedite the development process. Future commits may reintroduce this job or implement an alternative testing strategy for the frontend.